Updated change log consumer.
The server piece was re-written using Twisted.  I re-worked the protocol as a
line-oriented protocol.  I addressed a number of concerns the simple socket
server implementation had:

* The socket server worked fine as long as the network buffer sent all the
  data to the server in a single shot.  This could break under production
  conditions.
* The socket server used pickle/unpickle to serialize and unserialize the
  message that was transmitted.  This serialization format is not a
  particularly good choice when the data *may* come from an unknown source.
* The data travels over an unsecure transport.  Since the data itself is not
  sensitive and never leaves the server, this is acceptable.  However, an
  attacker that gained access to the server could send data to the server,
  so the message should be authenticated.  I implemented a simple HMAC
  scheme using a shared secret.

The client was modified to use the new server protocol.  Addtionally, it records
the last sequence number that it processed so it knows what point to start sending
messages from, even if it has been taken down while events occured.