Added contrib section for SysV Init scripts and supervisord configuration.
Since these scripts need to be run as daemon processes, init scripts are
a nice way to hook them into commands like `service`.  The `txgroupserver.py`
script can easilly be run as a daemon.  The `process_changelog.py` script
needs a little help, so using `supervisord` to manage it is recommended.