Revert "UNTESTED: Added the ability to transmit the message with an HMAC to prevent tampering."
This reverts commit 47afc32143ac38951109d5ec2281b56bf69678d7.

This idea does not work.  It does prevent tampering, and the lack of
confidentiality is not an issue per se, but without identity there is
no way to stop an attacker from replaying messages without adding
additional information to the messages (e.g. sequence numbers).

Eventually, this idea will turn into something almost entirely but
not quite unlike TLS.