Modify existing infrastructure to pass routing key and the unparsed message to the provisioner.
Only the provisioner will reasonably know what to do with the message.
The routing key may be useful information to provide to the provisioner, too.
For example, if the provisioner is an intermediate component, it could use
the remainder of the routing key to forward the message on to the next component
for processing.