Parsed messages now manage and serialize their own internals.
The messages just need help from the PSD with routing and possibly
attribute resolution.

Message parsers and parsed message formats will *probably* have a 1-to-1
correspondence, but keeping them separate still seems like a good idea.  E.g.
some other change log consumer could still parse into the same message format.

For the basic full sync, I'd like to keep it simple and not worry about
account provisioners.  The reason they are problematic is that either the
PSD needs to send a complete list of subject *and* their attributes
(seems totally inefficient!) OR the provisioner needs to compare the subject
list to what is really there and then send a request back to the PSD to resolve
attributes for a bunch of subjects.

At that point, the provisioner might just be better off talking to the
attribute resolver directly.

Membership full sync is just easier.