Changes to LDAP provisioner to support basic membership sync.
Made the decision to *not* allow normal updates and membership sync operations
to be interleaved.  This may cause some unnecessary waiting during a sync
operation (though not unresponsiveness).

The rationale is that if a group is updated while it is being synced,
a later update could be clobbered and lost, which is exactly the reason
you'd want to run a membership sync in the first place.

The argument against this is that updates for other groups may be delayed
unnecessarily.