BUGFIX: Removing a subject during workroom sync assumed access to the subject.
During a workroom sync, all the subjects that are supposed to be in the,
workroom are added, and then any that are in the workroom that aren't
supposed to be are removed.

During removal, the program only has a list of remote identifiers that
correspond to subjects-- *not* the local subject IDs.  The program can
make use of these remote identifiers to perform the task (in fact, they
are essential to the API).  However, for logging purposes, it is nice
to have the local identifiers.  Code assumed that it could easily retrieve
the local identifiers to use for logging, but this proved to be erroneous.